cambios pantallas y clase experto en proyecto

+++ sisexperto/sisexperto/Entidades/ExpertoEnProyecto.cs
@@ -22,7 +22,7 @@ namespace sisExperto.Entidades
         public virtual Experto Experto { get; set; }
 
         public double Ponderacion { get; set; }
-        public double Peso { get; set; }
+        public int Peso { get; set; }
 
         public virtual ValoracionCriteriosPorExperto ValoracionCriteriosPorExperto { get; set; }
         public virtual ICollection<ValoracionAlternativasPorCriterioExperto> ValoracionAlternativasPorCriterioExperto { get; set; }
